+++ src/config/common.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.1.0";
+export const VERSION = "0.2.0";
 export const CHANGELOG =
     "https://github.com/lajbel/kaplay/blob/master/CHANGELOG.md";
 export const REPO = "https://github.com/lajbel/kaplay";