+++ src/hooks/useUpdateEffect.ts
@@ -1,13 +1,13 @@
 import * as React from "react";
 
 export const useUpdateEffect: typeof React.useEffect = (effect, deps) => {
-	const isMountedRef = React.useRef(false);
+    const isMountedRef = React.useRef(false);
 
-	React.useEffect(() => {
-		if (!isMountedRef.current) {
-			isMountedRef.current = true;
-		} else {
-			return effect();
-		}
-	}, deps);
+    React.useEffect(() => {
+        if (!isMountedRef.current) {
+            isMountedRef.current = true;
+        } else {
+            return effect();
+        }
+    }, deps);
 };