+++ CHANGELOG.md
@@ -5,28 +5,35 @@ Format: Major.Minor.Patch (D/M/Y)
 ## 0.2.0 (unreleased)
 
 ### Features
+
 - ⭐ Multi-file editing support (with scenes)
 - added a loading screen
 - added reset project option
 - now editor is snapable at all
 
 ### Misc
+
 - new logo, hi dino!
 
 ## 0.1.1 (9/5/2024)
+
 - fixed code url not loading
 
 ## 0.1.0 (9/5/2024)
+
 The initial release!
 
 ### Features
+
 - added project import and export
 - added panes resizing
 - now assets dragging in editor put the asset in a new line
 
 ### Bug Fixes
+
 - fixed duplied assets
 - fixed (doubtful) a bug of editor writing
 
 ### Misc
+
 - added about the project window